Refactor: P1 pointers are now source indices
Previously, 'ptrs' was an array of line segments drawn from the source.
Now it is an array of indices which refer to source line segments. This
way, it is not required to know whether 'ptrs[p]' or 'p + t' is chosen
from the maximum statement on line 67; ptrs[p] can simply be
incremented to find the next possible match.
Comments were also cleaned up and now clearly refer to Ukkonen's
pseudocode from his 2003 paper.