Tests: New exact_matches and P2 tests
1) Added P1 option='segment' for exact_matches testing
2) Added a new test case to exact_matches testing called 'duplicate
melody'. This ensures that if a score has a duplicate occurrence of a
pattern (such as two voices singing the same line at the same time),
the algorithms will find all occurrences. P2 currently fails this test
and finds NO occurrences (because the way the shift multiplicity
counter works)
3) Added new test cases to test_P2 to ensure its approximate matching
works. These tests create random mismatches in the pattern (up to one
or to) and make sure the algorithm can still match approximately.