Fix: P1-3 pass all expected tests
Problem was that the midiparser was .ffill()-ing all of the pandas
dataframes, so none of the test data was actually good data. We put in
.ffill() to get measure information in June. So none of the testing has
been accurate since then; we've just been lucky. Good to know the
algorithms work though.
TODO: should put .ffill() just for measure information, and not for
duration nor noterest indexer.