Replaced "filter_occurrence" with "filter_result"
Previously, filter_occurrence only worked on a list of matching_pairs,
and algorithm output had to be translated to matching_pairs before it
could be filtered.
Filtering should happen as close as possible to the output of the
algorithm, since it's been policy to yield as many potential matches
as possible from them.
So now each alg has a custom filter_result, which is immediately called
outside of the algorithm and decides (using user settings) whether a
particular result is worth showing to the user.
This way, user settings remain separate from the algorithm code, but
filtering is still rather quick.