New: Setting validation and interval_func
Setting validation is now handled through the GeoAlgorithm class. Each
keyword which needs a special handling has a function with the name
_keyword, which either translated the keyword argument or raises an error if
the input was invalid.
Also added interval_func as a possible setting, which defines the
height between horizontal line segments (e.g. do we want enharmonic
equivalence? base40? generic vs chromatic intervals?). This function
is passed into InterNoteVector(), so the handling happens in the
algorithm class where the settings exist, not the interval object.