New: workflow and algorithm interface separation
Previously the GeoAlgorithm class acted as both an algorithm interface
for P, S, and W algorithms as well as a workflow (it was a generator
which parsed scores and filtered the results from the algorithm
function). Now, the workflow is separated in the Finder class, and the
GeoAlgorithm is an interface which demands a few functions from each
algorithm that inherits from it. these are:
pre_process (which takes NotePointSets and gets it ready for the
algorithm)
filter_result (which looks at user settings to ignore yielded results)
process_result (since algorithm output is not uniform, each one has to
implement its own result-to-occurrence translator)
and two functions which return generators; filtered_results and
processed_results, which apply filter_result() and process_result()
respectively on the algorithm output.

Now the user interfaces only with the Finder object, which can switch
and decide between the algorithms without having to re-parse the scores.