Refactor occurrences
Previously we just had one occurrence class. But each family of
algorithms will output different data (because they are operating
on different representations of music). So this should be reflected
in the occurrences they return.
Thus, we make a base occurrence class which is then subclassed by
more specific occurrences. The minimum amount of data for an occurrence
is the specific notes in the source which we are declaring as similar
to the query